(WHO-395) Implement Force HTTPS redirect
This adds the Force HTTPS redirect functionality. When enabled, accessing the
site over HTTP will trigger a 307 redirect the HTTPS version. The setting can
only be enabled when viewing the site over HTTPS in order to prevent users from
accidentally enabling the setting and locking themselves out. It also doesn't
trigger for the /healthz endpoint, since that needs to be accessible over HTTP
for load balancers to work.