(WHO-907) Simplify IndexFactoryProvider
This removes unused code from and simplifies `IndexFactoryProvider`,
specifically from the constructors. The `scaleFactor` parameter was unused, so
this removes it, as well as simplify down to a single constructor which
requires `dbPrefix` to be specified (even if `null`) so as to ensure that no
code accidentally forgets to set the `dbPrefix` when it should. This also sets
the constructor to `protected`, since nothing other than `EuropaInjectorModule`
should be calling it.