(WHO-1182) Model manual pipeline promotion
This updates the models for pipeline support to better handle manual promotion
gates. A `PipelineComponent` returns a richer object after `execute()` is
called, and `runPipeline` is updated to handle this data, which enables earlier
components to affect later components. This is important for adding support for
specifying an image tag to use at the time when you manually promote an image.