(WHO-1153) Save manifests more consistently
This updates the `RegistryManifestDb` so that when saving a registry manifest,
it first checks that there is already an entry for the raw manifest digest SHA
and, if not, saves a copy for that. This also removes similar code from the
`PCCopyToRepository` class, since it's not needed there anymore, as well as
removes the code there that creates a repo event since `RegistryManifestDb`
already does that.