(WHO-984) Improve _catalog pagination
This improves pagination for the _catalog endpoint by ensuring that the
requested number of entries, if there are that many, are returned, regardless
of access control. Without this, if a request lacks access to some of the
results, the number of results returned will be less than the full page size.

This also fixes the format of the `Link` headers returned to be in compliance
with RFC 5988.