(WHO-1286) Move Registry out of PCCopyToRepository
This (mostly) moves the `Registry` interface and its various implementations
out of `PCCopyToRepository` and into their own classes. It also starts moving
the logic on performing the copy operation into a separate class,
`CopyImageBetweenRepos`, in support of mirroring functionality, which will
require performing copy operations outside of a pipeline component.