(WHO-1286) Finish generalizing repo image copy
This finishes pulling the `Registry` code and capability for copying an image
between repositories out from `PCCopyToRepository` and into standalone classes,
in support of adding mirroring support. Specifically, this switches to using a
AssistedInject factories for `Registry` implementations in order that Guice can
actually inject things, and switches `PCCopyToRepository` to use the new
`CopyImageBetweenRepos` class, which it previously didn't.