(WHO-1264) add two new validation methods
First, there is no URL validation for the SAML page other than #validateNonNull. This adds a generalized URL validation method that tries to create a URL object and throws an Ajax exception if it cannot parse the URL string.
Second, there is no validation for the x509 certificate field other than the previously stated. This change adds a vaidation method that tries to create a certificate object and throws an Ajax exception if it cannot.