(WHO-1694) backend changes for auth token creation
Currently, token CRUD operations use the 'Requester' domain. So if you
have logged in as 'foo' (this is the Requester), you change the
context to the 'bar' (ctx or Owner) user, and create an Auth Token, the
token will be created as 'foo', not 'bar'.

This is a problem in the context of service accounts. The service
account should be able to create tokens such that access
is not based on the availability of a particluar engineer/developer
account.

Now, all tokens will be created using the Owner domain so that you can
CRUD tokens (based on set permissions) as a service account or your own.