(WHO-1779) add permissions checks for token operations
Most of the AJAX calls for tokens happen in europa.
Currently, there are no checks in any of them for permissions (because permissions didn't exist for auth tokens until now).
We are adding the ability to create tokens on other accounts so we want
to be able to lock that down with permissions. This change adds
permission checks to the CRUD classes for auth tokens.